requirements.txt
django-environ==0.8.1
django-extensions==3.1.5
django==4.0
djangorestframework==3.12.4
drf-extensions==0.7.1
pytz==2021.3